# .trae/documents/openapi-specifications.yaml
openapi: 3.0.3
info:
  title: RomAPI - Écosystème d'API Commercial Cameroun
  description: |
    API complète pour l'écosystème RomAPI qui centralise les ressources et services locaux du Cameroun.
    
    ## Authentification
    
    L'API utilise deux méthodes d'authentification :
    - **API Key** : Pour les intégrations tierces (header `X-API-Key`)
    - **Bearer Token** : Pour les utilisateurs authentifiés (header `Authorization: Bearer <token>`)
    
    ## Rate Limiting
    
    Les limites de taux dépendent de votre plan :
    - **Free** : 1,000 requêtes/mois
    - **Pro** : 50,000 requêtes/mois  
    - **Enterprise** : 500,000 requêtes/mois
    
    ## Formats Locaux
    
    - **Téléphone** : Format international (+237XXXXXXXXX) ou local (6XXXXXXXX)
    - **Adresse** : Formats camerounais avec régions et départements
    - **Fuseau horaire** : Africa/Douala (UTC+1)
    
  version: 1.0.0
  contact:
    name: RomAPI Support
    email: support@romapi.cm
    url: https://romapi.cm
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.romapi.cm/v1
    description: Production server
  - url: https://staging-api.romapi.cm/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /businesses:
    get:
      tags:
        - Businesses
      summary: Liste des entreprises
      description: Récupère la liste des entreprises avec filtres et pagination
      parameters:
        - name: page
          in: query
          description: Numéro de page (commence à 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Nombre d'éléments par page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: category
          in: query
          description: Filtrer par catégorie (ID ou slug)
          schema:
            type: string
            example: "restaurants"
        - name: city
          in: query
          description: Filtrer par ville
          schema:
            type: string
            example: "Douala"
        - name: region
          in: query
          description: Filtrer par région
          schema:
            type: string
            example: "Littoral"
        - name: lat
          in: query
          description: Latitude pour recherche géographique
          schema:
            type: number
            format: float
            example: 4.0511
        - name: lng
          in: query
          description: Longitude pour recherche géographique
          schema:
            type: number
            format: float
            example: 9.7679
        - name: radius
          in: query
          description: Rayon de recherche en km (avec lat/lng)
          schema:
            type: number
            format: float
            minimum: 0.1
            maximum: 100
            default: 10
        - name: plan
          in: query
          description: Filtrer par plan d'abonnement
          schema:
            type: string
            enum: [free, premium, featured]
        - name: verified
          in: query
          description: Filtrer par statut de vérification
          schema:
            type: boolean
        - name: sort
          in: query
          description: Critère de tri
          schema:
            type: string
            enum: [name, created_at, updated_at, distance, plan]
            default: plan
        - name: order
          in: query
          description: Ordre de tri
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Liste des entreprises
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Business'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'

    post:
      tags:
        - Businesses
      summary: Créer une entreprise
      description: Crée une nouvelle entreprise (authentification requise)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessCreate'
      responses:
        '201':
          description: Entreprise créée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Entreprise créée avec succès"
                  data:
                    $ref: '#/components/schemas/Business'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /businesses/{id}:
    get:
      tags:
        - Businesses
      summary: Détails d'une entreprise
      description: Récupère les détails complets d'une entreprise
      parameters:
        - name: id
          in: path
          required: true
          description: ID de l'entreprise
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails de l'entreprise
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BusinessDetailed'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimit'

    put:
      tags:
        - Businesses
      summary: Modifier une entreprise
      description: Modifie une entreprise existante (propriétaire uniquement)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID de l'entreprise
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessUpdate'
      responses:
        '200':
          description: Entreprise modifiée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Entreprise modifiée avec succès"
                  data:
                    $ref: '#/components/schemas/Business'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags:
        - Businesses
      summary: Supprimer une entreprise
      description: Supprime une entreprise (propriétaire uniquement)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID de l'entreprise
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entreprise supprimée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Entreprise supprimée avec succès"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /businesses/ingest:
    post:
      tags:
        - Businesses
      summary: Ingestion de données (n8n)
      description: Endpoint pour l'ingestion automatique de données via n8n
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: string
                  description: Source des données
                  example: "pages-jaunes-cm"
                source_url:
                  type: string
                  format: uri
                  description: URL source des données
                  example: "https://www.pagesjaunes.cm/restaurants-douala"
                businesses:
                  type: array
                  items:
                    $ref: '#/components/schemas/BusinessIngest'
      responses:
        '200':
          description: Données ingérées avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Données ingérées avec succès"
                  processed:
                    type: integer
                    example: 25
                  created:
                    type: integer
                    example: 20
                  updated:
                    type: integer
                    example: 5
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        business_name:
                          type: string
                        error:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  # Search Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Clé API pour l'authentification. Format: `romapi_live_xxxxxxxxxxxxxxxx`
        
        Obtenez votre clé API depuis votre dashboard utilisateur.
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via l'endpoint `/auth/login`.
        
        Format: `Authorization: Bearer <token>`
  schemas:
...
    Business:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Restaurant Le Palmier"
        slug:
          type: string
          example: "restaurant-le-palmier-douala"
        description:
          type: string
          example: "Restaurant traditionnel camerounais au cœur de Douala"
        category:
          $ref: '#/components/schemas/Category'
        address:
          $ref: '#/components/schemas/Address'
        contact:
          $ref: '#/components/schemas/Contact'
        status:
          type: string
          enum: [active, pending, suspended]
          example: "active"
        plan:
          type: string
          enum: [free, premium, featured]
          example: "premium"
        verified:
          type: boolean
          example: true
        primary_image:
          type: string
          format: uri
          example: "https://cdn.romapi.cm/images/business-123.jpg"
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.2
        reviews_count:
          type: integer
          example: 28
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-20T14:45:00Z"
    BusinessDetailed:
      allOf:
        - $ref: '#/components/schemas/Business'
        - type: object
          properties:
            hours:
              type: array
              items:
                $ref: '#/components/schemas/BusinessHour'
            images:
              type: array
              items:
                $ref: '#/components/schemas/BusinessImage'
            owner:
              type: object
              properties:
                name:
                  type: string
                  example: "Jean Dupont"
                verified:
                  type: boolean
                  example: true
            analytics:
              type: object
              properties:
                views_count:
                  type: integer
                  example: 1250
                clicks_count:
                  type: integer
                  example: 89
                contacts_count:
                  type: integer
                  example: 23
    BusinessCreate:
      type: object
      required:
        - name
        - category_id
        - address
        - contact
      properties:
        name:
          type: string
          example: "Restaurant Le Palmier"
        description:
          type: string
          example: "Restaurant traditionnel camerounais"
        category_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        address:
          $ref: '#/components/schemas/AddressCreate'
        contact:
          $ref: '#/components/schemas/ContactCreate'
        hours:
          type: array
          items:
            $ref: '#/components/schemas/BusinessHourCreate'
        images:
          type: array
          items:
            type: string
            format: uri
            example: "https://example.com/image.jpg"
    BusinessUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category_id:
          type: string
          format: uuid
        address:
          $ref: '#/components/schemas/AddressCreate'
        contact:
          $ref: '#/components/schemas/ContactCreate'
        hours:
          type: array
          items:
            $ref: '#/components/schemas/BusinessHourCreate'
    BusinessIngest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Restaurant Le Palmier"
        description:
          type: string
        category:
          type: string
          example: "Restaurants"
        address_line1:
          type: string
          example: "123 Rue de la Liberté"
        city:
          type: string
          example: "Douala"
        region:
          type: string
          example: "Littoral"
        phone:
          type: string
          example: "+237690123456"
        email:
          type: string
          format: email
        website:
          type: string
          format: uri
        source_id:
          type: string
          description: "ID unique dans la source"
          example: "pj_12345"
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Restaurants"
        slug:
          type: string
          example: "restaurants"
        description:
          type: string
          example: "Restaurants et établissements de restauration"
        icon:
          type: string
          example: "restaurant"
        parent_id:
          type: string
          format: uuid
          nullable: true
        businesses_count:
          type: integer
          example: 245
        created_at:
          type: string
          format: date-time
...
    Address:
      type: object
      properties:
        address_line1:
          type: string
          example: "123 Rue de la Liberté"
        address_line2:
          type: string
          example: "Quartier Bonanjo"
        city:
          type: string
          example: "Douala"
        region:
          type: string
          example: "Littoral"
        postal_code:
          type: string
          example: "1234"
        country:
          type: string
          example: "CM"
        latitude:
          type: number
          format: float
          example: 4.0511
        longitude:
          type: number
          format: float
          example: 9.7679
    AddressCreate:
      type: object
      required:
        - address_line1
        - city
        - region
      properties:
        address_line1:
          type: string
          example: "123 Rue de la Liberté"
        address_line2:
          type: string
          example: "Quartier Bonanjo"
        city:
          type: string
          example: "Douala"
        region:
          type: string
          example: "Littoral"
        postal_code:
          type: string
          example: "1234"
        country:
          type: string
          default: "CM"
          example: "CM"
    Contact:
      type: object
      properties:
        phone:
          type: string
          example: "+237690123456"
        email:
          type: string
          format: email
          example: "contact@restaurant-palmier.cm"
        website:
          type: string
          format: uri
          example: "https://restaurant-palmier.cm"
    ContactCreate:
      type: object
      properties:
        phone:
          type: string
          pattern: '^(\+237)?[0-9]{9}$'
          example: "+237690123456"
        email:
          type: string
          format: email
          example: "contact@restaurant-palmier.cm"
        website:
          type: string
          format: uri
          example: "https://restaurant-palmier.cm"
    BusinessHour:
      type: object
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: "0 = Dimanche, 1 = Lundi, ..., 6 = Samedi"
          example: 1
        open_time:
          type: string
          format: time
          example: "08:00"
        close_time:
          type: string
          format: time
          example: "22:00"
        is_closed:
          type: boolean
          example: false
    BusinessHourCreate:
      type: object
      required:
        - day_of_week
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
        open_time:
          type: string
          format: time
          example: "08:00"
        close_time:
          type: string
          format: time
          example: "22:00"
        is_closed:
          type: boolean
          default: false
...
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 156
        total_pages:
          type: integer
          example: 8
        has_next_page:
          type: boolean
          example: true
        has_prev_page:
          type: boolean
          example: false
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Les données fournies ne sont pas valides"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/v1/businesses"
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Bad Request"
            message: "Paramètres de requête invalides"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses"
    Unauthorized:
      description: Non autorisé - Token manquant ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Token d'authentification requis"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses"
    Forbidden:
      description: Accès interdit - Permissions insuffisantes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden"
            message: "Vous n'avez pas les permissions pour cette action"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses/123"
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not Found"
            message: "Entreprise non trouvée"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses/123"
    ValidationError:
      description: Erreur de validation des données
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: object
                    properties:
                      fields:
                        type: object
                        additionalProperties:
                          type: array
                          items:
                            type: string
          example:
            error: "Validation Error"
            message: "Les données fournies ne sont pas valides"
            details:
              fields:
                name: ["Le nom est requis"]
                email: ["Format d'email invalide"]
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses"
    RateLimit:
      description: Limite de taux dépassée
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: object
                    properties:
                      limit:
                        type: integer
                        example: 1000
                      remaining:
                        type: integer
                        example: 0
                      reset_at:
                        type: string
                        format: date-time
                        example: "2024-01-15T11:00:00Z"
          example:
            error: "Rate Limit Exceeded"
            message: "Limite de requêtes dépassée"
            details:
              limit: 1000
              remaining: 0
              reset_at: "2024-01-15T11:00:00Z"
            timestamp: "2024-01-15T10:30:00Z"
            path: "/v1/businesses"
...